Structure: Refactor hooks & GroupListItem
