+++ src/components/groups/groupListItem/index.tsx
